requirements.txt
alembic==1.6.5
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
bcrypt==3.2.0
black==21.7b0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
importlib-metadata==4.6.4
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
promise==2.3
psycopg2==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
regex==2021.8.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uvicorn==0.15.0
zipp==3.5.0
